Fix empty header
Some users have reported this issue on Rails 3.2.13:

    Net::SMTPFatalError
    (554 Transaction failed: Empty header names are illegal.)

The issue is that it is adding an extra new line to the original header,
which already ended with a new line character.
The extra empty line has been converted to a ':' which is illegal.

This fixes that issue. The specs on Ruby 2 were also printing this:

    WARNING: Could not parse (and so ignoring) ': '

With the fix, that warning is silenced.