Setting cc and bcc fields to nil instead of []
This was causing a problem with a certain `ActionMailer` provider
(Mailgun/Railgun).

This should also be the more canonical approach to un-setting these
fields as by default they are `nil`.

Setting the `Mail::Message`'s header differently was also required. The
previous approach would encode the message's headers to a string and
then append another encoded string to create the new headers. However,
the `bcc` field is not encoded by default
(see `Mail::BccField#encoded` @ 2.7.0).

Since the `bcc` field is not present in the encoded version of the
headers, it is effectively deleted and set to `[]`. Due to how setting
headers works in Mail, `message.bcc = nil` will not actually work.

By using `Mail::Header[]=` instead of `Mail::Header=`,
we do not clear the `bcc` field and can properly use
`message.bcc = nil` to remove it.