Disconnect Travis
GitHub is undergoing a migration from GitHub Services to GitHub Apps.

Travis is undergoing a migration from travis-ci.org to travis-ci.com.

It's more work to re-connect these things than it's worth
for the volume of change in this project.

Perhaps re-connect Travis in the future when things have settled.