Refactor
While preparing to add a new feature to this library
for the first time in years, I wanted to re-organize some of this code
so that I can better read it, understand it, and work with it.

There should be no functional changes here.
Mostly, I'm inlining code to reduce jumping around files
so it can be read more linearly.